requirements.txt
Django==1.11.6
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13
olefile==0.46
Pillow==4.3.0
PyJWT==1.6.4
pytz==2018.5
